+++ Net/Handlers/Helpers/MovementParser.cpp
@@ -20,7 +20,7 @@ std::vector<Movement> MovementParser::parse_movements(InPacket &recv) {
     std::vector<Movement> movements;
     uint8_t length = recv.read_byte();
 
-    for (uint8_t i = 0; i < length; ++i) {
+    for (int i = 0; i < length; ++i) {
         Movement fragment;
         fragment.command = recv.read_byte();
